requirements.txt
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
awsebcli==3.20.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blessed==1.19.0
botocore==1.21.65
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
cryptography==36.0.1
decorator==5.1.1
Django==4.0.1
django-extensions==3.1.5
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
executing==0.8.2
future==0.16.0
gunicorn==20.1.0
idna==3.3
ipython==8.0.0
jedi==0.18.1
jmespath==0.10.0
jsonschema==3.2.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
paramiko==2.9.2
parso==0.8.3
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
stack-data==0.1.4
termcolor==1.1.0
texttable==1.6.4
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.1.9
websocket-client==0.59.0
